Convert familySurnames to typescript
